fix: Move DatabaseName to data_types2
