+++ service_grpc_flight/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 # Workspace dependencies, in alphabetical order
-data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 datafusion = { path = "../datafusion" }
 generated_types = { path = "../generated_types" }
 observability_deps = { path = "../observability_deps" }
