+++ write_buffer/src/config.rs
@@ -193,7 +193,8 @@ mod tests {
         core::test_utils::random_topic_name, maybe_skip_kafka_integration,
         mock::MockBufferSharedState,
     };
-    use data_types::{write_buffer::WriteBufferCreationConfig, DatabaseName};
+    use data_types::write_buffer::WriteBufferCreationConfig;
+    use data_types2::DatabaseName;
     use std::{convert::TryFrom, num::NonZeroU32};
     use tempfile::TempDir;
